Now can search product and shop in product tab and list shop at account tab

+++ Shopee Prices Analystic/Models/Product.swift
@@ -12,5 +12,5 @@ struct Product {
     var name: String?
     var description: String?
     var price: Int?
-    var rating: Int?
+    var rating: Double?
 }
